requirements.txt
pyaudio
requests
openai
playsound
google-api-python-client
google-cloud-texttospeech